api/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
opensearch-py[async]>=2.4.0
redis>=5.0.0
PyJWT>=2.8.0
python-dateutil>=2.9.0
aiokafka>=0.11.0
